dls_pmaclib: Updated for compatibility with Clipper board.
Included Clipper as variant of Geobrick, as it has the same features.

Also corrected socket receive in _sendCommand for Ethernet interface. For
unknown reasons a SIGINT causes issues with the socket receive command
when opening the file browser for .pmc files, so socket receive is now
attempted multiple times if SIGINT blocks it.